feat: add new function to parse signed integers in lib.rs
A new function 'int<T>(input: &str)' has been implemented which can parse signed integers from a provided input string. This function employs the 'Neg' trait from std::ops and constructs an integer from a parsed unsigned one based on the presence of a leading '-'. This is a useful addition to the existing parser functionalities.